Remove the tpm2 from the single-encrypted-disk test
Reason: Beaker is having trouble provisioning machine
that has at least two disks and also has TPM2 module.

We were waiting a full day to get a machine and in the end
we have not even get the machine with tpm module.

Thus removing this case from the single encrypted disk test.